// src/models/profile/profile.interface.ts
export class Profile {
    uid?: string;
    firstName: string;
    lastName: string;
    email: string;
    dateOfBirth: any;
    job: string;
    location: string;
    avatar: string;
    cover: string;
    description: string;
    knowledgeNumber: number;
    knowledges: string;
    opinion: string;
    jobberStatut: boolean;
    userPermission : number;
    requestPost: number;
    workDone: number;
    commentPost: number;
    date: any;
}